requirements.txt
Flask>=2.2.2
Pillow>=9.2.0
qrcode>=7.3.1
requests>=2.25.1
six>=1.16.0
